requirements.txt
python-telegram-bot==20.7
pytz==2023.3
asyncio-compat==0.1.2
requests==2.31.0